added keploy tests
