+++ karbon/mocks/TestKarbonGetK8sRegistrationList.yaml
@@ -1,0 +1,54 @@
+version: api.keploy.io/v1beta2
+kind: Http
+name: TestKarbonGetK8sRegistrationList
+spec:
+    metadata:
+        Body: ""
+        Header: map[Accept:[application/json] Content-Type:[application/json] User-Agent:[nutanix]]
+        Proto: HTTP/1.1
+        ProtoMajor: "1"
+        ProtoMinor: "1"
+        URL: https://10.45.30.138:9440/karbon/v1-alpha.1/k8s/cluster-registrations/
+        name: http-client
+        operation: GET
+        type: HTTP_CLIENT
+    req:
+        method: GET
+        proto_major: 1
+        proto_minor: 1
+        url: https://10.45.30.138:9440/karbon/v1-alpha.1/k8s/cluster-registrations/
+        header:
+            Accept: application/json
+            Content-Type: application/json
+            User-Agent: nutanix
+        body: ""
+        body_type: utf-8
+    resp:
+        status_code: 200
+        header:
+            Content-Security-Policy: 'default-src ''self'' *.nutanix.com nutanix.com; script-src ''self'' ''unsafe-inline'' ''unsafe-eval''; style-src ''self'' ''unsafe-inline'' ''unsafe-eval''; connect-src ''self'' wss: https://downloads.frame.nutanix.com https://downloads.frame.nutanix.us; img-src ''self'' blob: data:; frame-src ''self'' https://*.nutanix.com blob: data:'
+            Content-Type: application/json
+            Date: Mon, 09 Jan 2023 19:30:13 GMT
+            Server: envoy
+            Set-Cookie: Expires=Mon, 09-Jan-2023 19:45:12 GMT;Path=/;Secure;HttpOnly;SameSite=Lax
+            Strict-Transport-Security: max-age=8640000:includeSubdomains
+            Vary: Accept-Encoding
+            X-Content-Type-Options: nosniff
+            X-Dns-Prefetch-Control: "off"
+            X-Envoy-Upstream-Service-Time: "1568"
+            X-Frame-Options: SAMEORIGIN
+            X-Ntnx-Env: pc
+            X-Ntnx-Product: pc.2021.9.0.4
+            X-Permitted-Cross-Domain-Policies: master-only
+            X-Powered-By: Express
+            X-Webkit-Csp: 'default-src ''self'' *.nutanix.com nutanix.com; script-src ''self'' ''unsafe-inline'' ''unsafe-eval'';  style-src ''self'' ''unsafe-inline'' ''unsafe-eval''; connect-src ''self'' wss: https://downloads.frame.nutanix.com https://downloads.frame.nutanix.us; img-src ''self'' blob: data:; frame-src ''self'' https://*.nutanix.com blob: data:'
+            X-Xss-Protection: 1; mode=block
+        body: |
+            [{"categories_mapping":{"kubernetes-cluster":"kubernetes-io-cluster-my-cluster"},"name":"my-cluster","status":"kActive","uuid":"eae7fe7e-34e8-4978-bb9a-e49157e858d6"},{"categories_mapping":{"KubernetesClusterName":"cluster1","KubernetesClusterUUID":"634F5852-CE1B-465C-A95A-9B8DFFBDDE42","kubernetes-io-cluster-cluster1":"owned"},"name":"cluster1","status":"kActive","uuid":"634f5852-ce1b-465c-a95a-9b8dffbdde42"},{"categories_mapping":{"kubernetes-cluster":"kubernetes-io-cluster-my-cluster"},"name":"my-cluster","status":"kActive","uuid":"eae7fe7e-34e8-4978-bb9a-e49157e858d5"}]
+        body_type: utf-8
+        status_message: ""
+        proto_major: 0
+        proto_minor: 0
+    objects:
+        - type: error
+          data: H4sIAAAAAAAA/wEAAP//AAAAAAAAAAA=
